+++ src/CSSList/CSSList.php
@@ -48,9 +48,9 @@ abstract class CSSList implements CSSElement, CSSListItem, Positionable
     protected $contents = [];
 
     /**
-     * @param int<0, max> $lineNumber
+     * @param int<1, max>|null $lineNumber
      */
-    public function __construct(int $lineNumber = 0)
+    public function __construct(?int $lineNumber = null)
     {
         $this->setPosition($lineNumber);
     }