+++ src/Position/Position.php
@@ -48,15 +48,14 @@ trait Position
     }
 
     /**
-     * @param int<0, max>|null $lineNumber
+     * @param int<1, max>|null $lineNumber
      * @param int<0, max>|null $columnNumber
      *
      * @return $this fluent interface
      */
     public function setPosition(?int $lineNumber, ?int $columnNumber = null): Positionable
     {
-        // The conditional is for backwards compatibility (backcompat); `0` will not be allowed in future.
-        $this->lineNumber = $lineNumber !== 0 ? $lineNumber : null;
+        $this->lineNumber = $lineNumber;
         $this->columnNumber = $columnNumber;
 
         return $this;