+++ src/Rule/Rule.php
@@ -44,10 +44,10 @@ class Rule implements Commentable, CSSElement, Positionable
 
     /**
      * @param non-empty-string $rule
-     * @param int<0, max> $lineNumber
-     * @param int<0, max> $columnNumber
+     * @param int<1, max>|null $lineNumber
+     * @param int<0, max>|null $columnNumber
      */
-    public function __construct(string $rule, int $lineNumber = 0, int $columnNumber = 0)
+    public function __construct(string $rule, ?int $lineNumber = null, ?int $columnNumber = null)
     {
         $this->rule = $rule;
         $this->setPosition($lineNumber, $columnNumber);