+++ src/Value/CSSFunction.php
@@ -27,9 +27,9 @@ class CSSFunction extends ValueList
      * @param non-empty-string $name
      * @param RuleValueList|array<Value|string> $arguments
      * @param non-empty-string $separator
-     * @param int<0, max> $lineNumber
+     * @param int<1, max>|null $lineNumber
      */
-    public function __construct(string $name, $arguments, string $separator = ',', int $lineNumber = 0)
+    public function __construct(string $name, $arguments, string $separator = ',', ?int $lineNumber = null)
     {
         if ($arguments instanceof RuleValueList) {
             $separator = $arguments->getListSeparator();